Refactor tests #383
