+++ tests/test_helpers.py
@@ -3,7 +3,7 @@ from open_mastr.utils.helpers import (
     validate_parameter_format_for_download_method,
     validate_parameter_format_for_mastr_init,
     validate_api_credentials,
-    transform_data_parameter,
+    transform_data_parameter, data_to_include_tables,
 )
 import pytest
 import sys
@@ -234,3 +234,24 @@ def test_transform_data_parameter():
 
 def test_validate_api_credentials():
     validate_api_credentials()
+
+
+def test_data_to_include_tables():
+    # Prepare
+    include_tables_list = [
+        "anlageneegwind",
+        "einheitenwind",
+        "anlageneegwasser",
+        "einheitenwasser",
+    ]
+    include_tables_str = ["einheitenstromverbraucher"]
+
+    map_to_db_table_list = ["market_actors", "market_roles"]
+    map_to_db_table_str = ["locations_extended"]
+
+
+    # Assert
+    assert include_tables_list == data_to_include_tables(data=["wind", "hydro"], mapping="write_xml")
+    assert include_tables_str == data_to_include_tables(data=["electricity_consumer"], mapping="write_xml")
+    assert map_to_db_table_list == data_to_include_tables(data=["market"], mapping="export_db_tables")
+    assert map_to_db_table_str == data_to_include_tables(data=["location"], mapping="export_db_tables")
